Implement session validation and refactor authentication logic for improved security
