+++ src/app/actions/auth.ts
@@ -4,10 +4,33 @@ import 'server-only';
 
 import { prisma } from '@/lib/db';
 import { LoginSchema, LoginState } from '@/lib/definitions';
-import { createSession, deleteSession } from '@/lib/session';
+import { createSession, decrypt, deleteSession } from '@/lib/session';
 import { Deluge } from '@ctrl/deluge';
 import { cookies } from 'next/headers';
 import { redirect } from 'next/navigation';
+
+export async function validateSession() {
+  const cookieStore = await cookies();
+  const token = cookieStore.get('deluge-next-session')?.value;
+
+  if (!token) {
+    return { authenticated: false };
+  }
+
+  const session = await decrypt(token);
+  if (!session?.sessionId) {
+    return { authenticated: false };
+  }
+
+  const sessionExists = await prisma.session.findUnique({
+    where: { id: session.sessionId },
+  });
+
+  return {
+    authenticated: !!sessionExists,
+    sessionId: sessionExists ? session.sessionId : null,
+  };
+}
 
 async function savePassword(password: string) {
   await prisma.password.create({
@@ -88,7 +111,7 @@ export async function login(
   redirect('/');
 }
 
-export async function signout() {
+export async function signOut() {
   const cookieStore = await cookies();
   const token = cookieStore.get('deluge-next-session')?.value;
