+++ src/app/api/deluge/torrent/files/route.ts
@@ -1,5 +1,7 @@
+import { validateSession } from '@/app/actions/auth';
 import { Deluge } from '@ctrl/deluge';
 import { NextRequest, NextResponse } from 'next/server';
+import 'server-only';
 import { z } from 'zod';
 
 const deluge = new Deluge({
@@ -16,6 +18,14 @@ const getTorrentFilesSchema = z.object({
 });
 
 export async function GET(request: NextRequest) {
+  const { authenticated } = await validateSession();
+  if (!authenticated) {
+    return NextResponse.json(
+      { success: false, error: 'Unauthorized' },
+      { status: 401 },
+    );
+  }
+
   const url = new URL(request.url);
   const torrentIdParam = url.searchParams.get('torrentId');
