+++ src/app/api/deluge/torrent/speed/route.ts
@@ -1,5 +1,7 @@
+import { validateSession } from '@/app/actions/auth';
 import { Deluge } from '@ctrl/deluge';
 import { NextRequest, NextResponse } from 'next/server';
+import 'server-only';
 import { z } from 'zod';
 
 const deluge = new Deluge({
@@ -21,6 +23,14 @@ const getTorrentSpeedResponseSchema = z.object({
 });
 
 export async function GET(request: NextRequest): Promise<NextResponse> {
+  const { authenticated } = await validateSession();
+  if (!authenticated) {
+    return NextResponse.json(
+      { success: false, error: 'Unauthorized' },
+      { status: 401 },
+    );
+  }
+
   const url = new URL(request.url);
   const torrentIdParam = url.searchParams.get('torrentId');
