+++ src/app/list/page.tsx
@@ -1,26 +1,13 @@
 import 'server-only';
 
+import { validateSession } from '@/app/actions/auth';
 import DelugePage from '@/app/list/components/deluge-page';
-import { prisma } from '@/lib/db';
-import { decrypt } from '@/lib/session';
-import { cookies } from 'next/headers';
 import { redirect } from 'next/navigation';
 
 export default async function Page() {
-  const cookieStore = await cookies();
-  const token = cookieStore.get('deluge-next-session')?.value;
+  const { authenticated } = await validateSession();
 
-  if (!token) {
-    redirect('/login');
-  }
-
-  const session = await decrypt(token);
-
-  const sessionExists = await prisma.session.findUnique({
-    where: { id: session?.sessionId },
-  });
-
-  if (!sessionExists) {
+  if (!authenticated) {
     redirect('/login');
   }
